+++ src/main/java/de/samply/frontend/FrontendConfiguration.java
@@ -3,6 +3,7 @@ package de.samply.frontend;
 import de.samply.app.ProjectManagerConst;
 import jakarta.annotation.PostConstruct;
 import lombok.Data;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
 
@@ -10,6 +11,7 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
 
+@Slf4j
 @Configuration
 @ConfigurationProperties(prefix = ProjectManagerConst.FRONTEND_CONFIG)
 @Data
@@ -23,6 +25,14 @@ public class FrontendConfiguration {
         Map<String, String> tempSites = new HashMap<>();
         sites.forEach((siteId, sitePath) -> tempSites.put(replaceHyphen(siteId), sitePath));
         sites = tempSites;
+        log.info("----------------------------------------");
+        log.info("Frontend Configuration:");
+        log.info("Base URL: {}", baseUrl);
+        log.info("Sites:");
+        sites.forEach((key, value) ->
+                log.info("- {} : {}", key, value)
+        );
+        log.info("----------------------------------------");
     }
 
     private String replaceHyphen(String var) {