+++ CHANGELOG.md
@@ -242,6 +242,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Form Repository
 - Fetch all project states depending on project code
 - Make user role thread safe
+- Log frontend config
 
 
 ### Changed